requirements.txt
gunicorn==20.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.0
whitenoise==5.3.0
pip==21.3.1
requests==2.26.0
psycopg2==2.9.2
dj-database-url==0.5.0
django-extensions==3.1.5
pyparsing==2.4.7
django-storages==1.12.3
django-utils-six==2.0
python-decouple==3.5
django-environ==0.8.1
dj-database-url==0.5.0
djangorestframework==3.13.1